requirements.txt
requests~=2.31.0
Pillow~=9.1.1
tqdm~=4.66.1
nltk~=3.8.1
accelerate==0.25.0
transformers==4.40.1
invisible_watermark==0.2.0
safetensors==0.4.1
diffusers==0.25.0
datasets~=2.18.0
aiohttp~=3.8.4
lpips==0.1.4
dists-pytorch==0.1
joblib~=1.3.2
pydantic~=2.5.3
matplotlib~=3.8.2
aioboto3~=12.4.0
stlpips_pytorch~=0.0.2
torchsummary~=1.5.1
google-cloud-aiplatform~=1.49.0
scikit-image~=0.23.2
dreamsim~=0.1.3
pyzmq~=25.1.2
stanza~=1.8.2
openai~=1.30.1
